Replaced // nothrow comments with BOOST_NOEXCEPT_OR_NOTHROW
rvalue-constructors triggered a warning "C26439 SPECIAL_NOEXCEPT" if code analysis was set to level "Microsoft Native Recommended Rules" when compiling with Visual Studio 2017. See https://docs.microsoft.com/en-us/visualstudio/code-quality/c26439?view=vs-2017

Also, it is better to use language constructs to represent semantics when possible. E.g. it is now possible to store objects of the affected classes in a vector and the vector no longer has to call the copy constructors when resizing its internal buffer. Not sure how relevant it is for these particular objects though.